+++ docs/sources/source/collections/nemo_nlp.rst
@@ -87,13 +87,13 @@ NLP Neural Modules
    :show-inheritance:
    :exclude-members: forward
    
-.. automodule:: nemo.collections.nlp.nm.trainables.dialogue_state_tracking.state_tracking_trade_nm
+.. automodule:: nemo.collections.nlp.nm.trainables.dialogue_state_tracking.trade_generator_nm
    :members:
    :undoc-members:
    :show-inheritance:
    :exclude-members: forward
 
-.. automodule:: nemo.collections.nlp.nm.trainables.joint_intent_slot.joint_intent_slot_nm
+.. automodule:: nemo.collections.nlp.nm.trainables.joint_intent_slot.joint_intent_slot_classifier_nm
    :members:
    :undoc-members:
    :show-inheritance:
