+++ nemo/collections/nlp/nm/trainables/dialogue_state_tracking/__init__.py
@@ -14,4 +14,4 @@
 # limitations under the License.
 # =============================================================================
 
-from nemo.collections.nlp.nm.trainables.dialogue_state_tracking.state_tracking_trade_nm import *
+from nemo.collections.nlp.nm.trainables.dialogue_state_tracking.trade_generator_nm import *
